requirements.txt
numpy==1.22.1
opencv-python==4.5.5.62
opencv-contrib-python==4.5.5.62
google-cloud-vision==2.6.3
python-dotenv==0.19.2
dotmap==1.3.30
namedivider-python==0.1.0
Japanera==0.1.1

# Labeling tool for items to be extracted by OCR.
labelImg==1.8.6

# Normalization of addresses.
kanjize==1.0.0
requests==2.25.1
pandas==1.2.4
cachetools==4.2.2